+++ __tests__/helpers.ts
@@ -11,6 +11,10 @@ import {
 import { protoTimestamp } from '../src/common/helpers';
 import { QueueItem } from '../src/connection/web-socket.connection';
 import { Character } from '../src/entities/character.entity';
+import {
+  DialogParticipant,
+  PreviousDialog,
+} from '../src/entities/continuation/previous_dialog.entity';
 import { InworldPacket, PacketId } from '../src/entities/inworld_packet.entity';
 import {
   ExtendedCapabilities,
@@ -70,6 +74,7 @@ export const generateSessionToken = () => Promise.resolve(session);
 
 export const capabilitiesProps: Capabilities = {
   audio: true,
+  continuation: true,
   emotions: true,
   interruptions: true,
   phonemes: true,
@@ -85,6 +90,7 @@ export const extendedCapabilitiesProps: ExtendedCapabilities = {
 
 export const extendedCapabilitiesRequestProps: ExtendedCapabilitiesRequest = {
   audio: true,
+  continuation: true,
   emotions: true,
   interruptions: true,
   phonemeInfo: true,
@@ -127,3 +133,19 @@ export const convertPacketFromProto = (proto: ProtoPacket) => {
 };
 
 export const extension = { convertPacketFromProto };
+
+export const phrases = [
+  {
+    talker: DialogParticipant.CHARACTER,
+    phrase: v4(),
+  },
+  {
+    talker: DialogParticipant.PLAYER,
+    phrase: v4(),
+  },
+  {
+    talker: DialogParticipant.UNKNOWN,
+    phrase: v4(),
+  },
+];
+export const previousDialog = new PreviousDialog(phrases);